Preserve un-ported Lua resource loader.
