+++ src/screens/Title.py
@@ -9,6 +9,62 @@ import pygame
 from .Base import Base, ColorFade
 
 
+'''
+local function loader(resource, file_list)
+    local yield = coroutine.yield
+
+    -- Load the files listed into the resource table.
+    for k,v in pairs(file_list.fonts) do
+        resource.fonts[k] = love.graphics.newFont(v.src, v.size)
+        if resource.fonts[k] == nil then
+            print('Unable to load font:', k)
+        end
+
+        yield(v.src)
+    end
+
+    for k,v in pairs(file_list.images) do
+        resource.images[k] = love.graphics.newImage(v)
+        if resource.images[k] == nil then
+            print('Unable to load image:', k)
+        end
+
+        yield(v)
+    end
+
+    for k,v in pairs(file_list.music) do
+        resource.music[k] = love.audio.newSource(v, 'stream')
+        if resource.music[k] == nil then
+            print('Unable to load music:', k)
+        end
+
+        yield(v)
+    end
+
+    for k,v in pairs(file_list.sounds) do
+        resource.sounds[k] = love.audio.newSource(v, 'static')
+        if resource.sounds[k] == nil then
+            print('Unable to load sound:', k)
+        end
+
+        yield(v)
+    end
+
+    for k, v in pairs(file_list.maps) do
+        resource.maps[k] = require(v)
+        if resource.maps[k] == nil then
+            print('Unable to load map:', k)
+        end
+
+        yield(v)
+    end
+
+    local title_text = resource.text:getText('title')
+    return title_text.loading_done
+end
+'''
+
+
 class Title(Base):
     def __init__(self, game):
         super().__init__(game)
@@ -64,3 +120,21 @@ class Title(Base):
 
     def update(self, dt):
         self.fade.update(dt)
+
+        if self.fade.isDone():
+            self.can_exit = True
+
+        '''
+        -- Load resources.
+        if self.loading_routine then
+            local resume = coroutine.resume
+            alive, self.loaded_resource = resume(self.loading_routine, self.resources, rsrc_list)
+            if not alive then
+                self.loading_finished = true
+                local title_text = self.resources.text:getText('title')
+                self.loaded_resource = title_text.loading_done
+            end
+        else
+            self.loading_routine = coroutine.create(loader)
+        end
+        '''
